update - added the extra components
added the Book component as well as its children, Title, Author, BookInfo and CoverImg. I mount each book by maping the books in the data array each as a Book component that get the data of the book as a prop. I then parse the relevant book data to each of the children from book, eg. the Title component gets the title data for the current book.